+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     - Updated `heapq` module support with changes in 3.x
     - Updated `csv` module support with changes in 3.x
 - Added/improved support for `__iter__` and `__call__` methods
+- Added support for `float.hex/fromhex`
 
 ### Fixed